+++ examples/cypher/src/packets.cpp
@@ -468,7 +468,9 @@ void feature_expansion_binary_state_control_long_form(
   State::Packet& packet, std::span<uint8_t const> bytes) {
   packet.desc_strs.back() += " - Binary State Control Long Form";
   packet.desc_strs.back() +=
-    std::format("\n- Address={}", (bytes[2uz] << 7u) | (bytes[1uz] & 0x7Fu));
+    std::format("\n- Address={}",
+                (static_cast<uint32_t>(bytes[2uz]) << 7u) |
+                  (static_cast<uint32_t>(bytes[1uz]) & 0x7Fu));
   packet.desc_strs.back() +=
     std::format("\n- State={}", bytes[1uz] & ztl::mask<7u> ? 1 : 0);
   packet.pattern_str += " 0 11000000 0 DLLLLLLL 0 HHHHHHHH";
@@ -504,8 +506,9 @@ void feature_expansion_time_and_date(State::Packet& packet,
       std::array buf{bytes[3uz], bytes[2uz]};
       memcpy(&f16, data(buf), sizeof(f16));
 #else
-      auto const f16{
-        dcc::float16_to_float32(bytes[2uz] << 8u | bytes[3uz] << 0u)};
+      auto const f16{dcc::float16_to_float32(
+        static_cast<uint16_t>(static_cast<uint32_t>(bytes[2uz]) << 8u |
+                              static_cast<uint32_t>(bytes[3uz]) << 0u))};
 #endif
       packet.desc_strs.back() +=
         std::format("\n- Time Scale={}", static_cast<float>(f16));
@@ -530,7 +533,8 @@ void feature_expansion_command_station_feature_identification(
   switch ([[maybe_unused]] auto const iiii{bytes[1uz] & 0x0Fu}) {
     case 0b1111u: {
       dcc::LocoFeatures const feats{
-        static_cast<uint16_t>(bytes[2uz] << 8u | bytes[3uz] << 0u)};
+        static_cast<uint16_t>(static_cast<uint32_t>(bytes[2uz]) << 8u |
+                              static_cast<uint32_t>(bytes[3uz]) << 0u)};
       packet.desc_strs.back() += "\n- Loco Features";
       packet.desc_strs.back() += std::format(
         "\n- Basic Addresses 100-127 as Extended={}",
@@ -580,7 +584,8 @@ void feature_expansion_command_station_feature_identification(
     }
     case 0b1110u: {
       dcc::AccessoryBroadcastFeatures const feats{
-        static_cast<uint16_t>(bytes[2uz] << 8u | bytes[3uz] << 0u)};
+        static_cast<uint16_t>(static_cast<uint32_t>(bytes[2uz]) << 8u |
+                              static_cast<uint32_t>(bytes[3uz]) << 0u)};
       packet.desc_strs.back() += "\n- Accessory and Broadcast Features";
       packet.desc_strs.back() += std::format(
         "\n- Addresses Offset by 4={}",
@@ -620,7 +625,8 @@ void feature_expansion_command_station_feature_identification(
     }
     case 0b1101u: {
       dcc::BiDiFeatures const feats{
-        static_cast<uint16_t>(bytes[2uz] << 8u | bytes[3uz] << 0u)};
+        static_cast<uint16_t>(static_cast<uint32_t>(bytes[2uz]) << 8u |
+                              static_cast<uint32_t>(bytes[3uz]) << 0u)};
       packet.desc_strs.back() += "\n- RailCom Features";
       packet.desc_strs.back() += std::format(
         "\n- RailCom={}",
@@ -810,8 +816,9 @@ void cv_access_xpom(State::Packet& packet, std::span<uint8_t const> bytes) {
   auto const ss{bytes[0uz] & 0b11u};
   auto const page_str{
     std::format("Page CV31={} CV32={}", bytes[1uz], bytes[2uz])};
-  auto const cv_addr{static_cast<uint32_t>(
-    bytes[1uz] << 16u | bytes[2uz] << 8u | bytes[3uz] << 0u)};
+  auto const cv_addr{static_cast<uint32_t>(bytes[1uz]) << 16u |
+                     static_cast<uint32_t>(bytes[2uz]) << 8u |
+                     static_cast<uint32_t>(bytes[3uz]) << 0u};
   packet.desc_strs.back() += std::format("\n- Sequence Number={:02b}", ss);
   switch (kk) {
     case 0b01u:
